# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Convention Cards Backend API
  version: 0.1.0
  description: Backend HTTP API for WBF-number login, player profiles, partnerships, convention cards, sharing, exports, and activity history.
servers:
  - url: http://localhost:3000
    description: Local API server
security:
  - bearerAuth: []
...
paths:
...
  /partnerships:
    get:
      tags: [Partnerships]
      summary: List partnerships for the signed-in player.
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Partnerships.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnershipListResponse"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Partnerships]
      summary: Create a pending partnership by partner WBF number.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [partnerWbfNumber]
              properties:
                partnerWbfNumber:
                  type: string
      responses:
        "201":
          description: Partnership created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnershipResponse"
        default:
          $ref: "#/components/responses/Error"
  /partnerships/{partnershipId}:
    parameters:
      - $ref: "#/components/parameters/PartnershipId"
    get:
      tags: [Partnerships]
      summary: Load a partnership visible to the signed-in player.
      responses:
        "200":
          description: Partnership.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnershipResponse"
        default:
          $ref: "#/components/responses/Error"
  /partnerships/{partnershipId}/approve:
    parameters:
      - $ref: "#/components/parameters/PartnershipId"
    post:
      tags: [Partnerships]
      summary: Approve an invitation as the partner.
      responses:
        "200":
          description: Partnership approved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnershipResponse"
        default:
          $ref: "#/components/responses/Error"
  /partnerships/{partnershipId}/decline:
    parameters:
      - $ref: "#/components/parameters/PartnershipId"
    post:
      tags: [Partnerships]
      summary: Decline an invitation as the partner.
      responses:
        "200":
          description: Partnership declined.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnershipResponse"
        default:
          $ref: "#/components/responses/Error"
  /partnerships/{partnershipId}/archive:
    parameters:
      - $ref: "#/components/parameters/PartnershipId"
    post:
      tags: [Partnerships]
      summary: Archive a partnership visible to the signed-in player.
      responses:
        "200":
          description: Partnership archived.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnershipResponse"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Supabase access token
  parameters:
...
    PartnershipId:
      name: partnershipId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    PartnershipStatus:
      type: string
      enum: [pending, approved, declined, archived]
...
    Partnership:
      type: object
      required: [id, ownerPlayerId, partnerPlayerId, partnerWbfNumber, status, approvedAt, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        ownerPlayerId:
          type: string
          format: uuid
        partnerPlayerId:
          type: [string, "null"]
          format: uuid
        partnerWbfNumber:
          type: string
        status:
          $ref: "#/components/schemas/PartnershipStatus"
        approvedAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PartnershipResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Partnership"
    PartnershipListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [partnerships]
          properties:
            partnerships:
              type: array
              items:
                $ref: "#/components/schemas/Partnership"
...